Rename package fabric-gui-v0 -> fabric-screen-v1
