+++ fabric-screen-v1/build.gradle
@@ -1,4 +1,4 @@
-archivesBaseName = "fabric-gui-v0"
+archivesBaseName = "fabric-screen-v1"
 version = getSubprojectVersion(project, "0.1.0")
 
 dependencies {
