+++ fabric-screen-v1/src/main/java/net/fabricmc/fabric/api/event/client/screen/ScreenInitCallback.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.fabricmc.fabric.api.event.client.gui;
+package net.fabricmc.fabric.api.event.client.screen;
 
 import java.util.List;
