+++ fabric-screen-v1/src/main/java/net/fabricmc/fabric/mixin/eventsscreen/MixinScreen.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.fabricmc.fabric.mixin.eventsgui;
+package net.fabricmc.fabric.mixin.eventsscreen;
 
 import java.util.List;
 
@@ -24,8 +24,8 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import net.fabricmc.fabric.api.event.client.gui.ScreenInitCallback;
-import net.fabricmc.fabric.api.event.client.gui.ScreenView;
+import net.fabricmc.fabric.api.event.client.screen.ScreenInitCallback;
+import net.fabricmc.fabric.api.event.client.screen.ScreenView;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.AbstractButtonWidget;
