+++ fabric-screen-v1/src/main/resources/fabric-screen-v1.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "net.fabricmc.fabric.mixin.eventsgui",
+  "package": "net.fabricmc.fabric.mixin.eventsscreen",
   "compatibilityLevel": "JAVA_8",
   "client": [
     "MixinScreen"
