+++ settings.gradle
@@ -19,7 +19,7 @@ include 'fabric-containers-v0'
 include 'fabric-content-registries-v0'
 include 'fabric-crash-report-info-v1'
 include 'fabric-events-interaction-v0'
-include 'fabric-gui-v0'
+include 'fabric-screen-v1'
 include 'fabric-events-lifecycle-v0'
 include 'fabric-item-groups-v0'
 include 'fabric-keybindings-v0'
